added functionality to currency conversion calculator. adjusted folder structure
