+++ src/components/index.js
@@ -2,3 +2,4 @@ export { default as NavBar } from "./NavBar";
 export { default as CoinCard } from "./CoinCard";
 export { default as BitcoinChart } from "./BitcoinChart";
 export { default as CoinList } from "./CoinList"
+export { default as CoinSummary } from "./CoinSummary"
